pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Dima Evan <dimaevan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
flake8 = "^4.0.1"
python-dotenv = "^0.19.2"
gunicorn = "^20.1.0"
django-bootstrap-v5 = "^1.0.7"
django-extensions = "^3.1.5"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
